moved a utility function to separate module so it can be imported elsewhere
